Can now scroll through staves in Rhythmic and Melodic Entries. This commit also includes the renaming of models -> utils
